Splits status condition files
