+++ internal/status/gatewayclassconditions_test.go
@@ -18,9 +18,6 @@ import (
 	"testing"
 	"time"
 
-	projectcontour "github.com/projectcontour/contour/apis/projectcontour/v1"
-	"github.com/projectcontour/contour/internal/k8s"
-
 	"github.com/stretchr/testify/assert"
 	apiequality "k8s.io/apimachinery/pkg/api/equality"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -28,43 +25,6 @@ import (
 	"k8s.io/apimachinery/pkg/util/validation/field"
 	gatewayapi_v1alpha1 "sigs.k8s.io/gateway-api/apis/v1alpha1"
 )
-
-func TestHTTPRouteAddCondition(t *testing.T) {
-
-	var testGeneration int64 = 7
-
-	simpleValidCondition := metav1.Condition{
-		Type:               string(gatewayapi_v1alpha1.ConditionRouteAdmitted),
-		Status:             projectcontour.ConditionTrue,
-		Reason:             "Valid",
-		Message:            "Valid HTTPRoute",
-		ObservedGeneration: testGeneration,
-	}
-
-	httpRouteUpdate := RouteConditionsUpdate{
-		FullName:   k8s.NamespacedNameFrom("test/test"),
-		Generation: testGeneration,
-		Conditions: make(map[gatewayapi_v1alpha1.RouteConditionType]metav1.Condition),
-	}
-
-	got := httpRouteUpdate.AddCondition(gatewayapi_v1alpha1.ConditionRouteAdmitted, metav1.ConditionTrue, "Valid", "Valid HTTPRoute")
-
-	assert.Equal(t, simpleValidCondition.Message, got.Message)
-	assert.Equal(t, simpleValidCondition.Reason, got.Reason)
-	assert.Equal(t, simpleValidCondition.Type, got.Type)
-	assert.Equal(t, simpleValidCondition.Status, got.Status)
-	assert.Equal(t, simpleValidCondition.ObservedGeneration, got.ObservedGeneration)
-}
-
-func newCondition(t string, status metav1.ConditionStatus, reason, msg string, lt time.Time) metav1.Condition {
-	return metav1.Condition{
-		Type:               t,
-		Status:             status,
-		Reason:             reason,
-		Message:            msg,
-		LastTransitionTime: metav1.NewTime(lt),
-	}
-}
 
 func TestComputeGatewayClassAdmittedCondition(t *testing.T) {
 	var errs field.ErrorList
